refactored onlineguru main-class, looked like a cave of badgers
