+++ onlineguru/src/main/java/no/ntnu/online/onlineguru/ShutdownThread.java
@@ -1,0 +1,34 @@
+package no.ntnu.online.onlineguru;
+
+import org.apache.log4j.Logger;
+
+public class ShutdownThread extends Thread {
+
+    private OnlineGuru onlineGuru = null;
+    private static Logger logger;
+
+    public ShutdownThread(OnlineGuru onlineGuru) {
+
+        super();
+        this.onlineGuru = onlineGuru;
+        logger = Logger.getLogger(this.getClass());
+    }
+
+    public void run() {
+
+        logger.info("[Shutdown thread] Shutting down");
+        onlineGuru.rudeDisconnect();
+
+        try {
+            logger.warn("[Shutdown thread] Giving Onlineguru 15 seconds to disconnect..");
+            sleep(15000L);
+        }
+        catch (InterruptedException e) {
+            logger.warn(e);
+        }
+
+        onlineGuru.stopThread();
+        logger.info("[Shutdown thread] Shutdown complete");
+    }
+
+}
